+++ all_consts.py
@@ -2,4 +2,4 @@ ALPHA = ["а","б","в","г","д","е","ё","ж","з","и","й",
          "к","л","м","н","о","п","р","с","т","у","ф",
          "х","ц","ч","ш","щ","ъ","ы","ь","э","ю","я"]
 MAX_ERRORS = 7
-JSON_WAY = "D:/backend/gallows/gallows.json"
+JSON_PATH = "D:/backend/gallows/gallows.json"
