+++ main.py
@@ -1,0 +1,4 @@
+from gallows import main
+
+if __name__ == "__main__":
+    main()
